@@ -6,41 +6,41 @@
         "name": "Floris Steenkamp"
     },
     "license": "MIT",
-    "main": "node/index.js",
-    "typings": "node/index",
+    "type": "module",
+    "main": "dist/index.js",
+    "engines": {
+        "node": "^12.20.0 || ^14.13.1 || >=16.0.0"
+    },
     "private": false,
     "dependencies": {
-        "@material-ui/core": "^4.11.0",
-        "@material-ui/icons": "^4.9.1",
-        "@types/ramda": "^0.27.14",
-        "@types/react-dom": "^16.9.8",
-        "@types/react-router-dom": "^5.1.5",
-        "@uirouter/angularjs": "^1.0.27",
-        "flo-bezier3": "^3.0.0",
-        "flo-boolean": "^1.0.0",
-        "flo-draw": "^1.0.0",
-        "flo-lines-intersections": "^1.0.1",
-        "flo-ll-rb-tree": "^0.1.2",
-        "flo-mat": "^1.2.0",
-        "flo-memoize": "^0.1.2",
-        "flo-poly": "^4.0.2",
-        "flo-vector2d": "^3.0.0",
-        "ramda": "^0.27.1",
-        "react": "^16.13.1",
-        "react-dom": "^16.13.1",
-        "react-router-dom": "^5.2.0",
-        "rxjs": "^6.6.2"
+        "@emotion/react": "^11.9.0",
+        "@emotion/styled": "^11.8.1",
+        "@mui/icons-material": "^5.6.2",
+        "@mui/material": "^5.7.0",
+        "@mui/styles": "^5.7.0",
+        "flo-bezier3": "^4.0.0",
+        "flo-boolean": "^2.0.0",
+        "flo-draw": "^2.0.0",
+        "flo-mat": "^2.0.0",
+        "flo-poly": "^6.0.4",
+        "flo-vector2d": "^4.0.1",
+        "ramda": "^0.28.0",
+        "react": "^17.0.2",
+        "react-dom": "^17.0.1",
+        "react-router-dom": "^6.3.0"
     },
     "devDependencies": {
-        "@stackblitz/sdk": "^1.5.0",
-        "@types/node": "^8.10.62",
-        "copy-webpack-plugin": "^6.0.3",
-        "source-map-loader": "^0.2.4",
-        "ts-loader": "^4.3.1",
-        "tsify": "^3.0.3",
-        "typescript": "^3.9.7",
-        "webpack": "^4.44.1",
-        "webpack-cli": "^3.3.12"
+        "@types/react": "^17.0.0",
+        "@types/react-dom": "^17.0.1",
+        "@types/node": "^17.0.34",
+        "@types/ramda": "^0.28.12",
+        "@types/react-router-dom": "^5.3.3",
+        "resolve-typescript-plugin": "^1.2.0",
+        "source-map-loader": "^3.0.1",
+        "ts-loader": "^9.3.0",
+        "typescript": "^4.6.4",
+        "webpack": "^5.72.1",
+        "webpack-cli": "^4.9.2"
     },
     "scripts": {
         "build": "webpack"
