@@ -1,5 +1,8 @@
+import { Theme } from "@mui/material";
 
-import { Theme, createStyles, makeStyles } from "@material-ui/core";
+
+import createStyles from '@mui/styles/createStyles';
+import makeStyles from '@mui/styles/makeStyles';
 
 
 const useStyles = makeStyles((theme: Theme) => createStyles({
