@@ -1,26 +1,19 @@
-
 declare var _debug_: Debug;
 
-import { Debug, toScaleAxis, Mat, simplifyMat, traverseEdges } from 'flo-mat';
-import { StateControl } from '../state-control/state-control';
+import { Debug, toScaleAxis, Mat, simplifyMat } from 'flo-mat';
 
 
 function getSatsFromGeneratedMats(
-        //stateControl: StateControl,
         mats: Mat[],
-        scale: number) {
+        scale: number): void {
 
-    //let { upd, state } = stateControl;
-    //let { appState } = state;
-    //let scale = appState.pageState.satScale;
-
+    if (_debug_ === undefined) { return; }
+    
     if (scale < 1) { return; }
 
     let satss: Mat[][] = [];
 
     let generated = _debug_.generated;
-
-    //let mats = generated.elems.mat;
 
     generated.elems.sat = [];
     generated.elems.maxVertex = [];
@@ -36,8 +29,6 @@ function getSatsFromGeneratedMats(
             return simplifyMat(toScaleAxis(mat, scale));
         })
     );
-
-    //console.log(satss)
 }
 
 
