@@ -1,10 +1,9 @@
-
 declare var _debug_: Debug; 
 
-import { Debug } from '../debug';
 import { squaredDistanceBetween } from 'flo-vector2d';
-import { closestPointOnBezierPrecise } from 'flo-bezier3';
+import { closestPointOnBezier } from 'flo-bezier3';
 import { drawFs } from 'flo-draw';
+import { Debug } from '../debug.js';
 
 
 function logNearestBezier(g: SVGGElement, p: number[], showDelay = 1000) {
@@ -19,7 +18,7 @@ function logNearestBezier(g: SVGGElement, p: number[], showDelay = 1000) {
         for (let loop of bezierLoops) {
             let beziers = loop.beziers;
             for (let ps of beziers) {
-                let bezierPoint = closestPointOnBezierPrecise(ps, p);
+                let bezierPoint = closestPointOnBezier(ps, p);
                 let d = squaredDistanceBetween(bezierPoint.p, p);
                 
                 if (d < bestDistance) {
