@@ -1,13 +1,12 @@
-
 declare var _debug_: Debug; 
 
-import { Debug } from '../debug';
 import { squaredDistanceBetween } from 'flo-vector2d';
-import { closestPointOnBezierPrecise } from 'flo-bezier3';
+import { closestPointOnBezier } from 'flo-bezier3';
 import { getLoopArea } from 'flo-boolean';
 import { beziersToSvgPathStr, Loop } from 'flo-mat';
-import { getLoopCentroid } from './get-loop-centroid';
 import { drawFs } from 'flo-draw';
+import { Debug } from '../debug.js';
+import { getLoopCentroid } from './get-loop-centroid.js';
 
 
 function logNearestLoopSimplified(g: SVGGElement, p: number[], showDelay = 1000) {
