@@ -1,5 +1,4 @@
-
-import { DebugElems } from '../debug';
+import { DebugElems } from '../debug.js';
 
 
 type ToDraw = { [T in keyof DebugElems]: boolean }
